requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
firebase-admin==6.5.0
python-dotenv==1.0.1
